Commit 26-03 Jens Fix???
