+++ .idea/libraries/org_openjfx_javafx_fxml_15_0_11.xml
@@ -1,0 +1,17 @@
+<component name="libraryTable">
+  <library name="org.openjfx:javafx-fxml:15.0.11" type="repository">
+    <properties maven-id="org.openjfx:javafx-fxml:15.0.1" />
+    <CLASSES>
+      <root url="jar://$PROJECT_DIR$/lib/javafx-fxml-15.0.1.jar!/" />
+      <root url="jar://$PROJECT_DIR$/lib/javafx-fxml-15.0.1-win.jar!/" />
+      <root url="jar://$PROJECT_DIR$/lib/javafx-controls-15.0.1.jar!/" />
+      <root url="jar://$PROJECT_DIR$/lib/javafx-controls-15.0.1-win.jar!/" />
+      <root url="jar://$PROJECT_DIR$/lib/javafx-graphics-15.0.1.jar!/" />
+      <root url="jar://$PROJECT_DIR$/lib/javafx-graphics-15.0.1-win.jar!/" />
+      <root url="jar://$PROJECT_DIR$/lib/javafx-base-15.0.1.jar!/" />
+      <root url="jar://$PROJECT_DIR$/lib/javafx-base-15.0.1-win.jar!/" />
+    </CLASSES>
+    <JAVADOC />
+    <SOURCES />
+  </library>
+</component>
