+++ Groepsproject.iml
@@ -4,10 +4,11 @@
     <exclude-output />
     <content url="file://$MODULE_DIR$">
       <sourceFolder url="file://$MODULE_DIR$" isTestSource="false" />
-      <excludeFolder url="file://$MODULE_DIR$/src" />
+      <sourceFolder url="file://$MODULE_DIR$/src" isTestSource="false" />
     </content>
     <orderEntry type="inheritedJdk" />
     <orderEntry type="sourceFolder" forTests="false" />
     <orderEntry type="library" name="org.openjfx:javafx-fxml:15.0.1" level="project" />
+    <orderEntry type="library" name="org.openjfx:javafx-fxml:15.0.11" level="project" />
   </component>
 </module>
