--- Magister200app.java
@@ -1,26 +1,0 @@
-import domeinLaag.Student;
-import javafx.application.Application;
-import javafx.fxml.FXMLLoader;
-import javafx.scene.Parent;
-import javafx.scene.Scene;
-import javafx.scene.control.TextField;
-import javafx.stage.Stage;
-
-import java.io.IOException;
-
-public class Magister200app extends Application {
-    public static void main(String[] args) {
-        launch(args);
-    }
-
-    @Override
-    public void start(Stage stage) throws Exception {
-        String fxmlPagina = "userInterfaceLaag/LoginScherm.fxml";
-        FXMLLoader loader = new FXMLLoader(getClass().getResource(fxmlPagina));
-        Parent root = loader.load();
-        stage.setTitle("Login scherm");
-        stage.setScene(new Scene(root));
-        stage.show();
-    }
-    Student s1 = new Student("Luca", 444,"luca.franken",0,100.00,0,"1234");
-}
