--- userInterfaceLaag/DocentenScherm.fxml
@@ -1,35 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-
-<?import javafx.scene.control.*?>
-<?import javafx.scene.layout.*?>
-<?import javafx.scene.shape.*?>
-<?import javafx.scene.text.*?>
-
-<AnchorPane prefHeight="500.0" prefWidth="800.0" xmlns="http://javafx.com/javafx/11.0.1" xmlns:fx="http://javafx.com/fxml/1" fx:controller="userInterfaceLaag.DocentenSchermController">
-   <children>
-      <Rectangle fx:id="rectangleBoven" arcHeight="5.0" arcWidth="5.0" fill="DODGERBLUE" height="32.0" stroke="BLACK" strokeType="INSIDE" width="800.0" />
-      <Button fx:id="loguitKnop" layoutX="733.0" layoutY="7.0" minHeight="17.0" minWidth="60.0" mnemonicParsing="false" onAction="#loguitEnAfsluiten" prefHeight="17.0" prefWidth="60.0" text="Loguit">
-         <font>
-            <Font size="9.0" />
-         </font></Button>
-      <Label fx:id="naamLabel" layoutX="500.0" layoutY="8.0" text="PLACEHOLDER VOOR NAAM" />
-      <TableView layoutX="251.0" layoutY="88.0" prefHeight="323.0" prefWidth="534.0">
-         <columns>
-            <TableColumn prefWidth="64.00000381469727" text="Naam" />
-            <TableColumn prefWidth="102.0" text="Studentnummer" />
-            <TableColumn prefWidth="156.59996032714844" text="E-mail" />
-            <TableColumn prefWidth="66.20001220703125" text="Roll Call" />
-            <TableColumn prefWidth="102.39996337890625" text="Aanwezigheid" />
-            <TableColumn prefWidth="41.20001220703125" text="Info" />
-         </columns>
-      </TableView>
-      <TableView layoutX="21.0" layoutY="88.0" prefHeight="323.0" prefWidth="186.0">
-         <columns>
-            <TableColumn prefWidth="68.19999504089355" text="Klas" />
-            <TableColumn prefWidth="58.400001525878906" text="Datum" />
-            <TableColumn prefWidth="58.0" text="Les" />
-         </columns>
-      </TableView>
-      <Button layoutX="579.0" layoutY="434.0" mnemonicParsing="false" text="Aanwijzigingen toepassen" />
-   </children>
-</AnchorPane>
