--- userInterfaceLaag/LeerlingHoofdscherm.fxml
@@ -1,29 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-
-<?import javafx.scene.chart.*?>
-<?import javafx.scene.control.*?>
-<?import javafx.scene.layout.*?>
-<?import javafx.scene.shape.*?>
-<?import javafx.scene.text.*?>
-
-<AnchorPane prefHeight="500.0" prefWidth="800.0" xmlns="http://javafx.com/javafx/11.0.1" xmlns:fx="http://javafx.com/fxml/1" fx:controller="userInterfaceLaag.LeerlingHoofdschermController">
-   <children>
-      <Rectangle fx:id="rectangleBoven" arcHeight="5.0" arcWidth="5.0" fill="DODGERBLUE" height="32.0" stroke="BLACK" strokeType="INSIDE" width="800.0" />
-      <Button fx:id="loguitKnop" cache="true" layoutX="733.0" layoutY="7.0" mnemonicParsing="false" onAction="#loguitEnSluiten" prefHeight="17.0" prefWidth="60.0" text="Loguit">
-         <font>
-            <Font size="9.0" />
-         </font>
-      </Button>
-      <Label fx:id="naamLabel" layoutX="500.0" layoutY="8.0" text="PLACEHOLDER VOOR NAAM" />
-      <TableView fx:id="aanwezigheidsTabel" layoutX="39.0" layoutY="82.0" prefHeight="343.0" prefWidth="450.0">
-        <columns>
-          <TableColumn prefWidth="101.19998931884766" text="Les" />
-          <TableColumn prefWidth="86.80001831054688" text="Datum" />
-            <TableColumn prefWidth="104.20001220703125" text="Docent" />
-            <TableColumn prefWidth="62.4000244140625" text="Lokaal" />
-            <TableColumn prefWidth="93.5999755859375" text="Aanwezigheid" />
-        </columns>
-      </TableView>
-      <PieChart fx:id="rollCallAttendance" layoutX="533.0" layoutY="109.0" prefHeight="136.0" prefWidth="200.0" title="Roll Call Attendance" />
-   </children>
-</AnchorPane>
